+++ src/functions.bash
@@ -3,6 +3,7 @@
 #
 
 function testmansh_run_linter() {
+  bl64_dbg_app_show_function "@"
   local container="$1"
   local format="$2"
   local report="$3"
@@ -76,37 +77,40 @@ function testmansh_run_linter() {
 }
 
 function testmansh_run_linter_container() {
+  bl64_dbg_app_show_function "@"
   local format="$1"
   local flags="$2"
-
-  bl64_cnt_setup || return $?
 
   shift
   shift
   # shellcheck disable=SC2086
   bl64_cnt_run_interactive \
     --volume "${TESTMANSH_PROJECT}:${TESTMANSH_CONTAINER64_PROJECT}" \
+    --workdir "${TESTMANSH_CONTAINER64_PROJECT}" \
     "${TESTMANSH_REGISTRY}/${TESTMANSH_IMAGES_LINT}" \
     $flags \
     "$@"
 }
 
 function testmansh_run_linter_native() {
+  bl64_dbg_app_show_function "@"
   local format="$1"
   local flags="$2"
-
-  bl64_check_command "$TESTMANSH_CMD_SHELLCHECK" || return $?
 
   shift
   shift
   cd "$TESTMANSH_PROJECT" || return 1
+  bl64_dbg_app_show_info "current path: $(pwd)"
+  bl64_dbg_app_trace_start
   # shellcheck disable=SC2086
   "$TESTMANSH_CMD_SHELLCHECK" \
     $flags \
     "$@"
+  bl64_dbg_app_trace_stop
 }
 
 function testmansh_run_test() {
+  bl64_dbg_app_show_function "@"
   local container="$1"
   local format="$2"
   local report="$3"
@@ -147,11 +151,11 @@ function testmansh_run_test() {
 }
 
 function testmansh_run_test_native() {
+  bl64_dbg_app_show_function "@"
   local report="$1"
   local flags="$2"
   local target="$3"
 
-  bl64_check_command "$TESTMANSH_CMD_BATS" || return $?
   bl64_dbg_app_show_vars 'target' 'flags'
 
   # shellcheck disable=SC2086
@@ -163,6 +167,7 @@ function testmansh_run_test_native() {
 }
 
 function testmansh_run_test_container() {
+  bl64_dbg_app_show_function "@"
   local report="$1"
   local flags="$2"
   local target="$3"
@@ -184,12 +189,11 @@ function testmansh_run_test_container() {
 }
 
 function testmansh_run_test_container_batscore() {
+  bl64_dbg_app_show_function "@"
   local container="$1"
   local target="$2"
   local flags="$3"
   local env_file="$4"
-
-  bl64_cnt_setup || return $?
 
   # shellcheck disable=SC2086
   bl64_cnt_run_interactive \
@@ -215,10 +219,9 @@ function testmansh_run_test_container_batscore() {
 }
 
 function testmansh_open_container() {
+  bl64_dbg_app_show_function
   local target='/bin/bash'
   local env_file=''
-
-  bl64_cnt_setup || return $?
 
   [[ -n "$TESTMANSH_ENV" && -r "$TESTMANSH_ENV" ]] && env_file="--env-file $TESTMANSH_ENV"
   bl64_dbg_app_show_vars 'env_file'
@@ -246,6 +249,7 @@ function testmansh_open_container() {
 }
 
 function testmansh_list_images() {
+  bl64_dbg_app_show_function
   bl64_msg_show_text "Container images for bats-core test cases:"
   echo "$TESTMANSH_IMAGES_TEST"
   bl64_msg_show_text "Container image for shellcheck:"
@@ -253,6 +257,7 @@ function testmansh_list_images() {
 }
 
 function testmansh_list_test_scope() {
+  bl64_dbg_app_show_function
   bl64_check_directory "$TESTMANSH_DEFAULT_TEST_PATH" 'default path for test-cases not found. Please use -c to indicate where the code is.' || return $?
   bl64_msg_show_text "Test cases scope for bats-core (project: $TESTMANSH_PROJECT)"
   # shellcheck disable=SC2164
@@ -261,6 +266,7 @@ function testmansh_list_test_scope() {
 }
 
 function testmansh_list_linter_scope() {
+  bl64_dbg_app_show_function
   bl64_check_directory "$TESTMANSH_DEFAULT_LINT_PATH" 'default path for source code files not found. Please use -c to indicate where cases are.' || return $?
   bl64_msg_show_text "Source code scope for shellcheck (project: $TESTMANSH_PROJECT)"
   # shellcheck disable=SC2164
@@ -269,16 +275,24 @@ function testmansh_list_linter_scope() {
 }
 
 function testmansh_initialize() {
-  local debug="$1"
-  local verbose="$2"
-  local container="$3"
-  local command="$4"
+  local command="$1"
+  local container="$2"
 
-  [[ -z "$command" ]] && testmansh_help && return 1
+  bl64_check_parameter 'command' ||
+    { testmansh_help && return 1; }
 
-  bl64_dbg_set_level "$debug" &&
-    bl64_msg_set_level "$verbose" ||
+  if [[ "$command" == 'open_container' ]]; then
+    bl64_check_parameter 'TESTMANSH_IMAGES_TEST' 'Please specify what container image to open with the parameter -e Image' ||
     return $?
+  elif [[ "$command" == 'run_linter' && "$container" == "$BL64_VAR_OFF" ]]; then
+    bl64_check_command "$TESTMANSH_CMD_SHELLCHECK" || return $?
+  elif [[ "$command" == 'run_test' && "$container" == "$BL64_VAR_OFF" ]]; then
+    bl64_check_command "$TESTMANSH_CMD_BATS" || return $?
+  fi
+
+  if [[ "$container" == "$BL64_VAR_ON" ]]; then
+    bl64_cnt_setup || return $?
+  fi
 
   TESTMANSH_PROJECT="${TESTMANSH_PROJECT:-$(pwd)}"
   TESTMANSH_DEFAULT_TEST_PATH="${TESTMANSH_PROJECT}/${TESTMANSH_DEFAULT_TEST_PREFIX}"
@@ -315,19 +329,21 @@ function testmansh_initialize() {
     TESTMANSH_CMD_BATS_HELPER_ASSERT="${TESTMANSH_CMD_BATS_HELPER_ASSERT:-/opt/bats-core/test_helpers/assert/load.bash}"
     TESTMANSH_CMD_BATS_HELPER_FILE="${TESTMANSH_CMD_BATS_HELPER_FILE:-/opt/bats-core/test_helpers/file/load.bash}"
   fi
-  bl64_dbg_app_show_info "[TESTMANSH_PROJECT_ROOT=${TESTMANSH_PROJECT_ROOT}]"
-  bl64_dbg_app_show_info "[TESTMANSH_PROJECT_BIN=${TESTMANSH_PROJECT_BIN}]"
-  bl64_dbg_app_show_info "[TESTMANSH_PROJECT_SRC=${TESTMANSH_PROJECT_SRC}]"
-  bl64_dbg_app_show_info "[TESTMANSH_PROJECT_LIB=${TESTMANSH_PROJECT_LIB}]"
-  bl64_dbg_app_show_info "[TESTMANSH_PROJECT_BUILD=${TESTMANSH_PROJECT_BUILD}]"
-  bl64_dbg_app_show_info "[TESTMANSH_TEST=${TESTMANSH_TEST}]"
-  bl64_dbg_app_show_info "[TESTMANSH_TEST_SAMPLES=${TESTMANSH_TEST_SAMPLES}]"
-  bl64_dbg_app_show_info "[TESTMANSH_TEST_LIB=${TESTMANSH_TEST_LIB}]"
-  bl64_dbg_app_show_info "[TESTMANSH_TEST_BATSCORE_SETUP=${TESTMANSH_TEST_BATSCORE_SETUP}]"
-  bl64_dbg_app_show_info "[TESTMANSH_CMD_BATS_HELPER_SUPPORT=${TESTMANSH_CMD_BATS_HELPER_SUPPORT}]"
-  bl64_dbg_app_show_info "[TESTMANSH_CMD_BATS_HELPER_ASSERT=${TESTMANSH_CMD_BATS_HELPER_ASSERT}]"
-  bl64_dbg_app_show_info "[TESTMANSH_CMD_BATS_HELPER_FILE=${TESTMANSH_CMD_BATS_HELPER_FILE}]"
+  bl64_dbg_app_show_vars \
+    'TESTMANSH_PROJECT_ROOT' \
+    'TESTMANSH_PROJECT_BIN' \
+    'TESTMANSH_PROJECT_SRC' \
+    'TESTMANSH_PROJECT_LIB' \
+    'TESTMANSH_PROJECT_BUILD' \
+    'TESTMANSH_TEST' \
+    'TESTMANSH_TEST_SAMPLES' \
+    'TESTMANSH_TEST_LIB' \
+    'TESTMANSH_TEST_BATSCORE_SETUP' \
+    'TESTMANSH_CMD_BATS_HELPER_SUPPORT' \
+    'TESTMANSH_CMD_BATS_HELPER_ASSERT' \
+    'TESTMANSH_CMD_BATS_HELPER_FILE'
 
+  return 0
 }
 
 function testmansh_help() {
@@ -358,7 +374,7 @@ The tool also sets and exports shell environment variables that can be used dire
   - TESTMANSH_CMD_BATS_HELPER_SUPPORT: full path to the bats-core support helper
   - TESTMANSH_CMD_BATS_HELPER_ASSERT: full path to the bats-core assert helper
   - TESTMANSH_CMD_BATS_HELPER_FILE: full path to the bats-core file helper' \
-  '
+    '
   -b           : Run bats-core tests
   -t           : Run shellcheck linter
   -q           : Open bats-core container